+++ src/lib/scryfall.js
@@ -8,6 +8,10 @@ function findCardByName (name) {
   })
 }
 
+function findCardByScryfallId (id) {
+  return scryfall.get(`/cards/${id}`)
+}
+
 function searchForCards (query) {
   return scryfall.get('/cards/search', {
     q: query
@@ -19,6 +23,7 @@ function formatCard (response, card = {}) {
   card.id = card.id || uuid()
   card.name = response.name
   card.typeLine = response.type_line
+  card.scryfallId = response.id
   card.manaCost = response.mana_cost || (response.card_faces && response.card_faces[0].mana_cost) || ''
   card.colorIdentity = response.color_identity
   card.price = {
@@ -37,6 +42,7 @@ function formatCard (response, card = {}) {
 
 module.exports = {
   findCardByName,
+  findCardByScryfallId,
   searchForCards,
   formatCard
 }